Refactor agent set imports and introduce AgentSetRegistry
- Updated import paths for AbstractAgentSet and AgentSetRegistry to reflect new module structure.
- Created a new concrete implementation of AgentSetRegistry in `agentsetregistry.py`, providing a collection for managing agent sets with DataFrame-based storage.
- Modified existing files to utilize the new AgentSetRegistry class, ensuring consistent usage across the codebase.
