+++ mesa_frames/concrete/agentsetregistry.py
@@ -53,7 +53,7 @@ from typing import Any, Literal, Self, cast, overload
 import numpy as np
 import polars as pl
 
-from mesa_frames.abstract.agents import AbstractAgentSetRegistry, AbstractAgentSet
+from mesa_frames.abstract.agentsetregistry import AbstractAgentSetRegistry, AbstractAgentSet
 from mesa_frames.types_ import (
     AgentMask,
     AgnosticAgentMask,
